chore: Core 및 utility 모듈 추가 및 domain 모듈 삭제
# Conflicts:
#	Projects/Domain/User/Sources/API/SignAPI.swift
#	Projects/Domain/User/Sources/API/SignRepository.swift
#	Projects/Domain/User/Sources/DTO/RequestDTO/SignRequestDTO.swift
#	Projects/Shared/ThirdPartyLips/Project.swift

(cherry picked from commit 902a8e6025279ff1e2fa2ac8928531f5a1ad9fc7)

# Conflicts:
#	Projects/Core/Utility/Interface/Dummy.swift
#	Projects/Core/Utility/Project.swift
#	Projects/Domain/Agency/Project.swift
#	Projects/Domain/Ledger/Project.swift
#	Projects/Domain/LedgerDetail/Project.swift
#	Projects/Domain/University/Project.swift
#	Projects/Domain/User/Sources/API/SignAPI.swift
#	Projects/Domain/User/Sources/API/SignRepository.swift
#	Projects/Domain/User/Sources/RequestDTO/SignModelRequestDTO.swift
#	Projects/Shared/ThirdPartyLips/Project.swift
