+++ Projects/Core/Core/Tests/CoreTests.swift
@@ -1,0 +1,13 @@
+import XCTest
+@testable import Core
+
+final class CoreTests: XCTestCase {
+
+    override func setUpWithError() throws {}
+
+    override func tearDownWithError() throws {}
+
+    func testExample() throws {
+        XCTAssertEqual(1, 1)
+    }
+}
