+++ Projects/Core/Utility/Demo/Sources/UtilityViewController.swift
@@ -1,0 +1,15 @@
+import UIKit
+
+final class UtilityViewController: UIViewController {
+    init() {
+        super.init(nibName: nil, bundle: nil)
+    }
+    @available(*, unavailable)
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+    }
+}
