+++ Projects/Core/Utility/Project.swift
@@ -1,23 +1,20 @@
 import ProjectDescription
 
 let project = Project(
-    name: "BaseFeature",
+    name: "Utility",
     options: .options(
       disableBundleAccessors: true,
       disableSynthesizedResourceAccessors: true
     ),
     targets: [
         Target(
-            name: "BaseFeature",
+            name: "Utility",
             platform: .iOS,
             product: .framework,
-            bundleId: "com.framework.moneymong.BaseFeature",
+            bundleId: "com.framework.moneymong.Utility",
             deploymentTarget: .iOS(targetVersion: "15.0", devices: .iphone),
             sources: ["Sources/**"],
             dependencies: [
-              .project(target: "DesignSystem", path: .relativeToRoot("Projects/Shared/DesignSystem")),
-              .project(target: "ThirdPartyLips", path: .relativeToRoot("Projects/Shared/ThirdPartyLips")),
-              .project(target: "Core", path: .relativeToRoot("Projects/Core/Core"))
             ]
         )
     ]
