+++ Projects/Core/Utility/Tests/UtilityTests.swift
@@ -1,0 +1,13 @@
+import XCTest
+@testable import Utility
+
+final class UtilityTests: XCTestCase {
+
+    override func setUpWithError() throws {}
+
+    override func tearDownWithError() throws {}
+
+    func testExample() throws {
+        XCTAssertEqual(1, 1)
+    }
+}
