--- Projects/Domain/Agency/Tests/AgencyDomainTests.swift
@@ -1,13 +1,0 @@
-import XCTest
-@testable import AgencyDomain
-
-final class AgencyDomainTests: XCTestCase {
-
-    override func setUpWithError() throws {}
-
-    override func tearDownWithError() throws {}
-
-    func testExample() throws {
-        XCTAssertEqual(1, 1)
-    }
-}
