--- Projects/Domain/Ledger/Tests/LedgerDomainTests.swift
@@ -1,13 +1,0 @@
-import XCTest
-@testable import LedgerDomain
-
-final class LedgerDomainTests: XCTestCase {
-
-    override func setUpWithError() throws {}
-
-    override func tearDownWithError() throws {}
-
-    func testExample() throws {
-        XCTAssertEqual(1, 1)
-    }
-}
