--- Projects/Domain/LedgerDetail/Tests/LedgerDetailDomainTests.swift
@@ -1,13 +1,0 @@
-import XCTest
-@testable import LedgerDetailDomain
-
-final class LedgerDetailDomainTests: XCTestCase {
-
-    override func setUpWithError() throws {}
-
-    override func tearDownWithError() throws {}
-
-    func testExample() throws {
-        XCTAssertEqual(1, 1)
-    }
-}
