--- Projects/Domain/University/Demo/Sources/UniversityDomainViewController.swift
@@ -1,15 +1,0 @@
-import UIKit
-
-final class UniversityDomainViewController: UIViewController {
-    init() {
-        super.init(nibName: nil, bundle: nil)
-    }
-    @available(*, unavailable)
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-    }
-}
