--- Projects/Domain/University/Sources/Dummy.swift
@@ -1,1 +1,0 @@
-// Dummy
