--- Projects/Domain/University/Tests/UniversityDomainTests.swift
@@ -1,13 +1,0 @@
-import XCTest
-@testable import UniversityDomain
-
-final class UniversityDomainTests: XCTestCase {
-
-    override func setUpWithError() throws {}
-
-    override func tearDownWithError() throws {}
-
-    func testExample() throws {
-        XCTAssertEqual(1, 1)
-    }
-}
