--- Projects/Domain/User/Demo/Sources/UserDomainViewController.swift
@@ -1,15 +1,0 @@
-import UIKit
-
-final class UserDomainViewController: UIViewController {
-    init() {
-        super.init(nibName: nil, bundle: nil)
-    }
-    @available(*, unavailable)
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-    }
-}
