--- Projects/Domain/User/Interface/Dummy.swift
@@ -1,1 +1,0 @@
-// Dummy
