--- Projects/Domain/User/Sources/RequestDTO/SignModelRequestDTO.swift
@@ -1,6 +1,0 @@
-import Foundation
-
-public struct SignModelRequestDTO: Encodable {
-  let provider: String
-  let accessToken: String
-}
