--- Projects/Domain/User/Tests/UserDomainTests.swift
@@ -1,13 +1,0 @@
-import XCTest
-@testable import UserDomain
-
-final class UserDomainTests: XCTestCase {
-
-    override func setUpWithError() throws {}
-
-    override func tearDownWithError() throws {}
-
-    func testExample() throws {
-        XCTAssertEqual(1, 1)
-    }
-}
