--- Projects/Shared/Utils/Demo/Sources/UtilsViewController.swift
@@ -1,15 +1,0 @@
-import UIKit
-
-final class UtilsViewController: UIViewController {
-    init() {
-        super.init(nibName: nil, bundle: nil)
-    }
-    @available(*, unavailable)
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-    }
-}
