--- Projects/Shared/Utils/Project.swift
@@ -1,21 +1,0 @@
-import ProjectDescription
-
-let project = Project(
-  name: "Utils",
-  options: .options(
-    disableBundleAccessors: true,
-    disableSynthesizedResourceAccessors: true
-  ),
-  targets: [
-    Target(
-      name: "Utils",
-      platform: .iOS,
-      product: .framework,
-      bundleId: "com.framework.moneymong.Utils",
-      deploymentTarget: .iOS(targetVersion: "15.0", devices: .iphone),
-      sources: ["Sources/**"],
-      dependencies: [
-      ]
-    )
-  ]
-)
