--- Projects/Shared/Utils/Sources/Dummy.swift
@@ -1,1 +1,0 @@
-// Dummy
