--- Projects/Shared/Utils/Tests/UtilsTests.swift
@@ -1,13 +1,0 @@
-import XCTest
-@testable import Utils
-
-final class UtilsTests: XCTestCase {
-
-    override func setUpWithError() throws {}
-
-    override func tearDownWithError() throws {}
-
-    func testExample() throws {
-        XCTAssertEqual(1, 1)
-    }
-}
